+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.0.0] - 2021-09-06
+- Bug fixes
+
+## [0.1.3] - 2021-04-05
+### Removed
+- Removed LIBPAX_MAX_SIZE from API, since value is now fixed to 0xFFFF
+- Redesgined hash table (now using a bitmap) to improve memory efficiency and lookup speed
 
 ## [0.1.2] - 2021-03-23
 ### Removed